Socket assembly schematic
Functional-test socket design with important dimensions of the socket indicated. Only three of the four flexible struts are depicted. The socket can accommodate variable residual limb lengths by trimming the struts to the desired length. The struts can also be heated and reshaped to adapt to the residual limb shape of a given user. The struts are nested inside of a collet which in turn houses a custom adapter for a terminal device. The adapter for the TASKA hand is pictured here.